Modify: fix bugs from Main.java and MainWindow.java for jar file to work properly
