+++ src/main/java/zirnitra/ui/DialogBox.java
@@ -1,7 +1,5 @@
 package zirnitra.ui;
 
-import java.io.IOException;
-import java.util.Collections;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
@@ -12,6 +10,9 @@ import javafx.scene.control.Label;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.HBox;
+
+import java.io.IOException;
+import java.util.Collections;
 
 /**
  * A custom dialog box used in the Zirnitra GUI.
@@ -33,7 +34,8 @@ public class DialogBox extends HBox {
      */
     private DialogBox(String text, Image img) {
         try {
-            FXMLLoader fxmlLoader = new FXMLLoader(ZirnitraGui.class.getResource(FXML_PATH));
+            FXMLLoader fxmlLoader = new FXMLLoader(MainWindow.class
+                    .getResource(FXML_PATH));
             fxmlLoader.setController(this);
             fxmlLoader.setRoot(this);
             fxmlLoader.load();
